Windows fixes

+++ spec/worker.js
@@ -15,27 +15,37 @@ const wait = util.promisify(setTimeout);
 const EXPECTED_CALL_COUNT = 3;
 
 if (isMainThread) {
-  module.exports = function spawnThread(index) {
+  module.exports = function spawnThread(index, indexOfEarlyReturn) {
     let id = index + 1;
     return new Promise(async (resolve, reject) => {
       console.log('Spawning worker:', id);
       const worker = new Worker(__filename, {
-        workerData: id,
+        workerData: { id, earlyReturn: indexOfEarlyReturn === null ? false : id === (indexOfEarlyReturn + 1) },
       });
       worker.on('message', async (msg) => {
         console.log('[parent] Worker', id, 'reported call count:', msg);
-        await wait(500);
-        if (msg >= EXPECTED_CALL_COUNT) {
+        await wait(1000);
+        let expected = id === indexOfEarlyReturn + 1 ? (EXPECTED_CALL_COUNT - 1) : EXPECTED_CALL_COUNT;
+        let passes = msg >= expected;
+        if (passes) {
+          console.log(`Worker ${id} passed!`);
           resolve();
         } else {
-          reject(`Not enough calls! Expected: ${EXPECTED_CALL_COUNT} Actual: ${msg}`);
+          reject(`Not enough calls on worker ${id}! Expected: ${expected} Actual: ${msg}`);
         }
       });
-      worker.on('error', reject);
+      worker.on('error', (err) => {
+        console.error(`ERROR IN WORKER: ${id}`);
+        console.error(err);
+        reject(err);
+      });
       worker.on('exit', (code) => {
         if (code !== 0) {
-          console.log(`Worker stopped with exit code ${code}`);
+          console.log(`Worker ${id} stopped with exit code ${code}`);
           reject();
+        } else {
+          console.log(`Worker ${id} exited gracefully`);
+          // resolve();
         }
       });
     });
@@ -45,6 +55,8 @@ if (isMainThread) {
   const tempFile = path.join(tempDir, 'file');
 
   const { watch, closeAllWatchers } = require('../src/main');
+
+  console.log('NEW WORKER:', workerData);
 
   class Scheduler {
     constructor(id, pathToWatch) {
@@ -69,30 +81,36 @@ if (isMainThread) {
   }
 
   (async () => {
-    console.log('Worker', workerData, 'creating file:', tempFile);
+    console.log('Worker', workerData.id, 'creating file:', tempFile);
     fs.writeFileSync(tempFile, '');
     await wait(500);
-    const scheduler = new Scheduler(workerData, tempFile);
+    const scheduler = new Scheduler(workerData.id, tempFile);
     scheduler.start();
     await wait(2000);
 
-    console.log('Worker', workerData, 'changing file:', tempFile);
+    console.log('Worker', scheduler.id, 'changing file:', tempFile);
     // Should generate one or two events:
     fs.writeFileSync(tempFile, 'changed');
     await wait(1000);
-    console.log('Worker', workerData, 'changing file again:', tempFile);
+    console.log('Worker', scheduler.id, 'changing file again:', tempFile);
     // Should generate another event:
     fs.writeFileSync(tempFile, 'changed again');
-    await wait(1000);
-    // Should generate a final event (total count 3 or 4):
-    console.log('Worker', workerData, 'deleting file:', tempFile);
-    fs.rmSync(tempFile);
-    await wait(1000);
+    if (workerData.earlyReturn) {
+      console.log('Worker', scheduler.id, 'returning early!');
+    } else {
+      await wait(1000);
+      // Should generate a final event (total count 3 or 4):
+      console.log('Worker', scheduler.id, 'deleting file:', tempFile);
+      fs.rmSync(tempFile);
+      await wait(1000);
+
+      await wait(Math.random() * 2000);
+    }
 
     parentPort.postMessage(scheduler.callCount);
 
     closeAllWatchers();
-    console.log('Worker', workerData, 'closing');
-    process.exit(0);
+    console.log('Worker', scheduler.id, 'closing');
+    // process.exit(0);
   })();
 }
